list: reorganize utils tests
