+++ change/@fluentui-react-42e62350-c4bd-4d7e-bf80-eb79ef462001.json
@@ -1,6 +1,6 @@
 {
   "type": "patch",
-  "comment": "feat: list add utilities for reading and writing scroll values",
+  "comment": "feat: add list utilities for reading and writing scroll values",
   "packageName": "@fluentui/react",
   "email": "seanmonahan@microsoft.com",
   "dependentChangeType": "patch"
